+++ test/stream.js
@@ -1,9 +1,18 @@
 var Qqwry = require('../');
-var lib = Qqwry(true);
+var qqwry = Qqwry();
 
-var ss = lib.searchIPScopeStream('220.255.255.255', '255.255.255.255', { type: 'csv', outHeader: true });
+// var ss = lib.searchIPScopeStream('223.0.0.0', '255.255.255.255', { type: 'csv', outHeader: true });
+// ss.pipe(process.stdout);
 
-ss.pipe(process.stdout);
+qqwry.searchIPScopeStream("8.8.8.0","8.8.8.0").pipe(process.stdout);
+// qqwry.searchIPScopeStream("8.8.8.0","8.8.8.8",{format:'csv'}).pipe(process.stdout);
+// qqwry.searchIPScopeStream("8.8.8.0","8.8.8.8",{format:'json',outHeader:true}).pipe(process.stdout);
+// var ss = qqwry.searchIPScopeStream("8.8.8.0","8.8.8.8",{format:'object'});
+// ss.pipe(process.stdout);
+// ss.on('data',function(data){
+// 	console.log('> ', typeof data);
+// })
+
 
 // while (true) {
 // 	var obj = ss.read();